Search event view works properly
