+++ app/src/main/res/layout/activity_event_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.belatrixsf.connect.ui.contacts.ContactsListActivity">
+    tools:context="com.belatrixsf.connect.ui.event.EventListActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
