+++ app/src/main/res/layout/activity_user.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".ui.home.MainActivity">
+    tools:context=".ui.home.UserActivity">
 
     <android.support.design.widget.CoordinatorLayout
         android:id="@+id/main_coordinator"
